requirements-testing.txt
mock
pytest-cache
pytest-cov
pytest-django
pytest-pep8
